+++ contracts/interfaces/ITransferSelectorNFT.sol
@@ -7,13 +7,13 @@ pragma solidity ^0.8.17;
  */
 
 interface ITransferSelectorNFT {
-    event NewAssetType(uint8 assetType, address transferManager, bytes4 selector);
+    event NewAssetType(uint256 assetType, address transferManager, bytes4 selector);
 
     // Custom errors
     error AlreadySet();
-    error NFTTransferFail(address collection, uint8 assetType);
-    error NoTransferManagerForAssetType(uint8 assetType);
-    error WrongAssetType(uint8 assetType);
+    error NFTTransferFail(address collection, uint256 assetType);
+    error NoTransferManagerForAssetType(uint256 assetType);
+    error WrongAssetType(uint256 assetType);
 
     // Custom structs
     struct ManagerSelector {
